@@ -1,2 +1,7 @@
-'use strict';
-module.exports = x => typeof x === 'string' ? x.replace(/[\\"]/g, '\\$&') : x;
+export default function escapeStringAppleScript(string) {
+	if (typeof string !== 'string') {
+		throw new TypeError(`Expected a string, got ${typeof string}`);
+	}
+
+	return string.replace(/[\\"]/g, '\\$&');
+}
